Tests/LibWeb/Text/expected/css/logical-border-radius-computed.txt

border-start-start-radius: 50% -> 50%
border-start-end-radius: 25% -> 25%
border-end-start-radius: 10% -> 10%
border-end-end-radius: 75% -> 75%
border-start-start-radius: calc(10px + 5px) -> 15px
border-start-end-radius: calc(50% - 10%) -> 40%
border-end-start-radius: calc(20px * 2) -> 40px
border-end-end-radius: calc(100% / 4) -> 25%
